requirements.txt
alembic==1.15.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autopep8==2.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
build==1.2.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
check50==3.3.11
clang-format==18.1.7
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
docopt==0.6.2
EditorConfig==0.12.4
executing==2.0.1
faiss-cpu==1.10.0
fastjsonschema==2.20.0
ffmpeg==1.4
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
fqdn==1.5.1
fsspec==2025.3.2
gcloud==0.18.3
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.30.1
icdiff==2.0.7
idna==3.10
importlib_metadata==8.0.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jellyfish==0.11.2
Jinja2==3.1.5
joblib==1.4.2
jsbeautifier==1.15.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lib50==3.0.11
linker==0.1.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.2.2
oauth2client==4.1.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
php==1.2.1
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2025.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rank-bm25==0.2.2
redis==5.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.33.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
setuptools==78.1.0
setuptools-metadata==0.1.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sox==1.5.0
spotipy==2.24.0
SQLAlchemy==2.0.40
stack-data==0.6.3
style50==2.10.3
submit50==3.1.5
sympy==1.13.1
tbb==2021.13.0
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tokenizers==0.15.2
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.38.2
trio==0.30.0
trio-websocket==0.12.2
turtle==0.0.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.13.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uv==0.7.13
virtualenv==20.27.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.11
wsproto==1.2.0
yarg==0.1.9
zipp==3.19.2
